Add support for drawPicture rasterization strategy
