+++ packages/vector_graphics/lib/vector_graphics_compat.dart
@@ -1,0 +1,15 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'src/vector_graphics.dart'
+    show
+        AssetBytesLoader,
+        NetworkBytesLoader,
+        PictureInfo,
+        BytesLoader,
+        VectorGraphic,
+        VectorGraphicUtilities,
+        vg,
+        RenderingStrategy,
+        createCompatVectorGraphic;
