+++ packages/vector_graphics/test/render_vector_graphics_test.dart
@@ -11,7 +11,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:vector_graphics/src/listener.dart';
-import 'package:vector_graphics/src/render_vector_graphics.dart';
+import 'package:vector_graphics/src/render_vector_graphic.dart';
 import 'package:vector_graphics_codec/vector_graphics_codec.dart';
 
 import 'caching_test.dart';
