+++ packages/vector_graphics/test/vector_graphics_test.dart
@@ -9,8 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:vector_graphics/src/listener.dart';
-import 'package:vector_graphics/vector_graphics.dart';
-
+import 'package:vector_graphics/src/vector_graphics.dart';
 import 'package:vector_graphics_codec/vector_graphics_codec.dart';
 
 const VectorGraphicsCodec codec = VectorGraphicsCodec();
@@ -461,6 +460,31 @@ void main() {
         isA<OpacityLayer>()
             .having((OpacityLayer layer) => layer.alpha, 'alpha', 128));
   }, skip: !kIsWeb);
+
+  testWidgets('Construct vector graphic with drawPicture strategy',
+      (WidgetTester tester) async {
+    final TestAssetBundle testBundle = TestAssetBundle();
+
+    await tester.pumpWidget(
+      DefaultAssetBundle(
+        bundle: testBundle,
+        child: Directionality(
+          textDirection: TextDirection.ltr,
+          child: createCompatVectorGraphic(
+            loader: const AssetBytesLoader('foo.svg'),
+            colorFilter: const ColorFilter.mode(Colors.red, BlendMode.srcIn),
+            opacity: const AlwaysStoppedAnimation<double>(0.5),
+          ),
+        ),
+      ),
+    );
+    await tester.pumpAndSettle();
+
+    expect(tester.layers.last, isA<PictureLayer>());
+    // Opacity and color filter are drawn as savelayer
+    expect(tester.layers, isNot(contains(isA<OpacityLayer>())));
+    expect(tester.layers, isNot(contains(isA<ColorFilterLayer>())));
+  });
 }
 
 class TestAssetBundle extends Fake implements AssetBundle {
